controller for default profile image

+++ controllers/playersController.js
@@ -53,6 +53,10 @@ export const playersController = {
                     public_id: imagePosted.public_id
                 }
                 await fs.remove(req.files.image.tempFilePath)
+            } else {
+                image = {
+                    url: 'https://res.cloudinary.com/dlah9v2do/image/upload/v1679335452/1200px-Breezeicons-actions-22-im-user.svg_ycuwsn.png'
+                }
             }
 
             const newPlayer = new Player({
@@ -119,13 +123,15 @@ export const playersController = {
     },
     deletePlayer: async (req, res) => {
         try {
-            const playerRemoved = await Player.findByIdAndDelete(req.params.id);
-
-            if (playerRemoved.image.public_id) {
+            const playerRemoved = await Player.findById(req.params.id);
+            
+            if (playerRemoved.image.public_id && playerRemoved.image.url !== 'https://res.cloudinary.com/dlah9v2do/image/upload/v1679335452/1200px-Breezeicons-actions-22-im-user.svg_ycuwsn.png') {
                 await deleteImage(playerRemoved.image.public_id)
             }
+
+            await Player.deleteOne({_id: playerRemoved._id})
             
-            res.status(204)
+            res.status(204).send()
         } catch (error) {
             res.status(409).json({
                 message: error.message
